+++ src/Canvas/Background/Background.php
@@ -2,11 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Yilanboy\Preview\Contracts;
+namespace Yilanboy\Preview\Canvas\Background;
 
 use GdImage;
 
 interface Background
 {
+
     public function draw(GdImage $image, int $width, int $height): void;
+
 }